Separate build.yml into pr.yml and push.yml ++ Badges
By having the build.yml into two files, we can run different actions in pull requests and when pushing to main. This way we can add many more things (for example: coverage in OpenShift) in the push.yml to verify everything works as expected without having pull requests to wait all these actions to finish. 

Moreover, this is necessary to having a badge to see the health status of the project because there will be only one push action at a time, when it could be many pull requests actions (the badge will pick the latest one of a pull request that is not merged yet). 

Fix https://github.com/halkyonio/primaza-poc/issues/19